refactor: 클래스명, package 변경 (WinningCounter > WinningResult)
